requirements.txt
asgiref==3.3.1
astroid==2.4.2
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
Django==3.1.7
djangorestframework==3.12.2
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.20.0
pdfminer3==2018.12.3.0
pycryptodome==3.9.9
pylint==2.6.0
PyPDF2==1.26.0
pyTelegramBotAPI==3.7.4
pytz==2021.1
requests==2.25.1
schedule==1.0.0
six==1.15.0
sortedcontainers==2.3.0
soupsieve==2.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1